requirements.txt
altgraph==0.17.3
Pillow==9.3.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
tk==0.1.0
